@@ -6,11 +6,11 @@
   "dependencies": {
     "classnames": "^2.2.6",
     "object-assign": "^4.1.1",
-    "prop-types": "^15.6.2",
-    "react": "^16.8.1",
-    "react-dom": "^16.8.1",
+    "prop-types": "^15.7.2",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
     "react-proxy": "^1.1.8",
-    "uglifyjs-webpack-plugin": "^2.1.1"
+    "uglifyjs-webpack-plugin": "^2.2.0"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
