@@ -5,50 +5,42 @@
 import React from "react";
 import FacebookIcon from "./FacebookIcon";
 
-
-export default class FacebookLoginButton extends React.Component {
-
-  constructor(props) {
-    super(props);
-  }
-
+class FacebookLoginButton extends React.Component {
   /**
    *
    * @constructor
    */
-  FBLoginDialog() {
-
+  FBLoginDialog = () => {
     this.props.onStartLoading();
 
-    FB.login((response) => {
-
-      if (response.status == "connected" && this.props.onSuccess) {
-        this.props.onSuccess('facebook', response);
-
-      } else if (response.status != "connected" && this.props.onFail) {
-
-        this.props.onFail('facebook', response);
-      }
-    }, {scope: this.props.scope});
-  }
+    FB.login(
+      response => {
+        if (response.status == "connected" && this.props.onSuccess) {
+          this.props.onSuccess("facebook", response);
+        } else if (response.status != "connected" && this.props.onFail) {
+          this.props.onFail("facebook", response);
+        }
+      },
+      { scope: this.props.scope }
+    );
+  };
 
   /**
    *
    * @constructor
    */
-  render () {
-
+  render() {
     return (
       <button
         className={this.props.btnClass}
         disabled={this.props.inactive}
-        onClick={() => this.FBLoginDialog()}
+        onClick={this.FBLoginDialog}
       >
         <FacebookIcon />
-        <span>
-          {this.props.label}
-        </span>
+        <span>{this.props.label}</span>
       </button>
-    )
+    );
   }
-};
+}
+
+export default FacebookLoginButton;
