@@ -4,8 +4,7 @@
 
 import React from "react";
 
-export default class FormSubmitButton extends React.Component {
-
+class FormSubmitButton extends React.Component {
   constructor(props) {
     super(props);
   }
@@ -15,16 +14,17 @@ export default class FormSubmitButton extends React.Component {
    * @constructor
    */
   render() {
-
     return (
       <button
         className={this.props.buttonClass}
         disabled={this.props.inactive}
-        onClick={() => this.props.click()}
+        onClick={this.props.click}
         id={this.props.type + "Submit"}
       >
         {this.props.label}
       </button>
-    )
+    );
   }
-};
+}
+
+export default FormSubmitButton;
